pyproject.toml
[tool.poetry]
name = "django-poetry"
version = "0.1.0"
description = ""
authors = ["Anibal Sanchez <anibalnuma@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
phonenumbers = "^8.13.45"
graphene-django = "^3.2.2"
channels = "^4.1.0"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pylint = "^3.2.7"
pylint-django = "^2.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
